refactor to jsx
